備考を修正&&ダイアログをページに変更
